refactor the code .
